% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abertura.R
\name{get_abertura}
\alias{get_abertura}
\title{Captura o paragrafo inicial da ata}
\usage{
get_abertura(html_doc, possui_ata_complementar = NA)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{possui_ata_complementar}{boolean indicando se o pregao possui ata complementar
complementar ou nao; se NA, a funcao ira verificar no html da ata}
}
\value{
String com o paragrafo inicial da ata
}
\description{
Captura o paragrafo inicial da ata
}
