% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bid_data.R
\name{get_bid_data}
\alias{get_bid_data}
\title{Capturar dados do pregao}
\usage{
get_bid_data(html_doc, eventos = FALSE, mensagens = FALSE,
  return_item_id = TRUE, item_info = c("indice_item", "descricao",
  "descricao_complementar", "quantidade", "valor_estimado",
  "unidade_fornecimento", "tratamento_diferenciado", "decreto_7174",
  "margem_preferencia", "situacao", "resultado", "propostas", "lances",
  "fases"), indices = "all")
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{eventos}{se TRUE, eventos do pregao serao capturados}

\item{mensagens}{se TRUE, mensagens do pregao serao capturadas}

\item{item_info}{vetor de strings contendo as informacoes que serao capturadas, podendo incluir: 'indice_id', 'descricao', 'descricao_complementar', 'quantidade', 'valor_estimado', 'unidade_fornecimento', 'tratamento_diferenciado', 'decreto_7174', 'margem_preferencia', 'situacao', 'resultado', 'propostas', 'lances', 'lances_desempate_ME_EPP', 'eventos_item' e 'fases'}

\item{indices}{vetor numerico indicando os itens a serem capturados, caso nao se deseje capturar todos; a opcao default captura todos os itens do pregao}
}
\value{
Retorna um uma tibble, com estrutura de dataframe aninhado. Os dados referentes aos itens do pregao ficam na coluna 'dados', em um novo dataframe. Para cada item desse segundo dataframe, dados como historico de lances e propostas sao incluidos em uma terceira camada de dataframes aninhados.
}
\description{
Principal funcao para capturar os dados disponiveis na ata de um pregao eletronico do Comprasnet
}
