% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bid_events.R
\name{get_bid_events}
\alias{get_bid_events}
\title{Captura os eventos registrados da ata do pregao}
\usage{
get_bid_events(html_doc)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}
}
\value{
Retorna um dataframe com informacoes sobre os eventos registrados na ata
}
\description{
Captura os eventos registrados da ata do pregao
}
