% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bid_info.R
\name{get_bid_info}
\alias{get_bid_info}
\title{Captura informacoes do pregao que nao se referem aos itens negociados}
\usage{
get_bid_info(html_doc, eventos = FALSE, mensagens = FALSE)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{eventos}{se TRUE, eventos do pregao serao capturados}

\item{mensagens}{se TRUE, mensagens do pregao serao capturadas}
}
\value{
dataframe com as informacoes
}
\description{
Funcao utilizada para capturar informacoes do pregao que sao invariantes no que diz respeito aos itens
}
\seealso{
Para capturar informacoes referentes a itens de uma ata, ver get_item_info
}
