% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bid_info2.R
\name{get_bid_info2}
\alias{get_bid_info2}
\title{Captura informacoes do pregao que nao se referem aos itens negociados}
\usage{
get_bid_info2(html_doc, info = c("id_pregao", "data_abertura",
  "data_encerramento"))
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{eventos}{se TRUE, eventos do pregao serao capturados}

\item{mensagens}{se TRUE, mensagens do pregao serao capturadas}
}
\value{
dataframe com as informacoes
}
\description{
Funcao utilizada para capturar informacoes do pregao que sao invariantes no que diz respeito aos itens
}
\seealso{
Para capturar informacoes referentes a itens de uma ata, ver get_item_info
}
