% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bid_messages.R
\name{get_bid_messages}
\alias{get_bid_messages}
\title{Captura as mensagens registradas na ata do pregao}
\usage{
get_bid_messages(html_doc)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}
}
\value{
dataframe com registro das trocas de mensagens
}
\description{
Captura as mensagens registradas na ata do pregao
}
