% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_couasg.R
\name{get_couasg}
\alias{get_couasg}
\title{Captura codigo da UASG}
\usage{
get_couasg(html_doc, possui_ata_complementar = NA)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{possui_ata_complementar}{boolean indicando se o pregao possui ata complementar
complementar ou nao; se NA, a funcao ira verificar no html da ata}
}
\value{
String de 6 digitos com o codigo da UASG
}
\description{
Captura codigo da UASG
}
