% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Captura data e horario de abertura do pregao}
\usage{
get_data(x)
}
\arguments{
\item{x}{character string contendo o primeiro paragrafo ou xml_document da ata}
}
\value{
character string com data e hora de abertura do pregao
}
\description{
Captura data e hora de abertura do pregao, a partir do html da ata ou de string contendo o primeiro paragrafo da ata
}
