% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_encerramento.R
\name{get_encerramento}
\alias{get_encerramento}
\title{Captura o paragrafo final da ata}
\usage{
get_encerramento(html_doc)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}
}
\value{
String com o paragrafo final da ata
}
\description{
Captura o paragrafo final da ata
}
