% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id_pregao.R
\name{get_id_pregao}
\alias{get_id_pregao}
\title{Captura o numero identificador do pregao (16 digitos)}
\usage{
get_id_pregao(html_doc, possui_ata_complementar = NA)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{possui_ata_complementar}{boolean indicando se o pregao possui ata complementar
complementar ou nao; se NA, a funcao ira verificar no html da ata}
}
\value{
string com o numero identificador de 16 digitos
}
\description{
Funcao para capturar o id do pregao. O id do pregao possui 16 digitos, sendo os 6 primeiros correspondentes ao codigo da UASG e os 10 ultimos referentes ao numero do pregao (com zero-padding)
}
