% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_events.R
\name{get_item_events}
\alias{get_item_events}
\title{Captura o registro de eventos especificos ao item}
\usage{
get_item_events(html_doc, indice_item)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{indice_item}{string com o numero do item}
}
\value{
dataframe com o registro de eventos do item
}
\description{
Captura o registro de eventos especificos ao item
}
