% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_info_chr.R
\name{get_item_info_chr}
\alias{get_item_info_chr}
\title{Captura uma informacao especifica referente a um item do pregao}
\usage{
get_item_info_chr(item_node, item_info)
}
\arguments{
\item{item_node}{node do item cujos dados serao extraidos}

\item{item_info}{string indicando qual informacao se deseja capturar}
}
\value{
string com a informacao desejada, caso encontrada
}
\description{
Funcao para capturar uma informacao especifica e unidimensional (string) referente a um item.
}
\seealso{
para capturar dados referentes a itens disponiveis em formato de tabelas (e.g, historico de lances e propostas), ver funcao get_item_info_dfs
}
