% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_info_dfs.R
\name{get_item_info_df}
\alias{get_item_info_df}
\title{Captura informacoes referentes a itens que sao apresentadas no formato de tabelas}
\usage{
get_item_info_df(html_doc, id_item, item_info)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico ou xml_node referente ao item identificado no argumento id_item}

\item{id_item}{string com o numero do item}

\item{item_info}{string com a informacao a ser capturada}
}
\value{
dataframe com as informacoes
}
\description{
Funcao utilizada para capturar as seguintes informacoes relativas a itens especificos:
\itemize{
\item propostas
\item lances
\item lances_desempate_ME_EPP
\item eventos_item
}
}
\seealso{
Para capturar informacoes referentes a itens de uma ata, ver get_bid_info_dfs
}
