% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_nodes.R
\name{get_item_nodes}
\alias{get_item_nodes}
\title{Identifica os xml nodes dos itens}
\usage{
get_item_nodes(html_doc, possui_ata_complementar = NA)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{possui_ata_complementar}{boolean indicando se o pregao possui ata complementar
complementar ou nao; se NA, a funcao ira verificar no html da ata}
}
\value{
nodeset com todos os nodes de itens listados no inicio da ata
}
\description{
Identifica os xml nodes dos itens
}
