% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lances.R
\name{get_lances}
\alias{get_lances}
\title{Captura historico de lances do item}
\usage{
get_lances(html_doc, item_id)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{item_id}{string com o numero do item}
}
\value{
dataframe com o historico de lances, contendo cnpj do fornecedor, valor do lance e horario em que o lance foi registrado. Em alguns casos, o dataframe tambem contem informacoes adicionais como desconto, valor com desconto, fator de equalizacao e/ou valor equalizado
}
\description{
Captura o historico de lances do item
}
