% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lances_clean.R
\name{get_lances_clean}
\alias{get_lances_clean}
\title{Captura historico de lances do item e limpa os dados}
\usage{
get_lances_clean(html_doc, indice_item, clean = TRUE)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{indice_item}{string com o numero do item}

\item{clean}{boolean}
}
\value{
dataframe com o historico de lances, contendo cnpj do fornecedor, valor do lance e horario em que o lance foi registrado. Em alguns casos, o dataframe tambem contem informacoes adicionais como desconto, valor com desconto, fator de equalizacao e/ou valor equalizado
}
\description{
Captura o historico de lances do item, com a opcao de limpar os dados automaticamente. Funcao em desenvolvimento.
}
