% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lances_desempate_ME_EPP.R
\name{get_lances_desempate_ME_EPP}
\alias{get_lances_desempate_ME_EPP}
\title{Captura registro de lances de desempate de microempresa ou empresa de pequeno porte}
\usage{
get_lances_desempate_ME_EPP(html_doc, indice_item)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{indice_item}{string com o numero do item}
}
\value{
dataframe com os lances de desempate ou string 'Sem lances de desempate'. Na maior parte dos pregoes, nao ha lances de desempate
}
\description{
Captura registro de lances de desempate de microempresa ou empresa de pequeno porte
}
