% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_melhor_lance.R
\name{get_melhor_lance}
\alias{get_melhor_lance}
\title{Extrai dado de melhor lance a partir do texto do campo 'resultado'}
\usage{
get_melhor_lance(resultado)
}
\arguments{
\item{resultado}{character string com o resultado de um item, tal como disponivel no comeco das atas}
}
\value{
numeric
}
\description{
Extrai dado de melhor lance a partir do texto do campo 'resultado'
}
\seealso{
para capturar texto de resultado a partir do html, ver get_item_info e get_item_info_chr
}
