% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_ata.R
\name{get_num_ata}
\alias{get_num_ata}
\title{Captura o numero de atas complementares que o pregao possui}
\usage{
get_num_ata(html_doc)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}
}
\value{
string com o numero de atas complementares do pregao
}
\description{
Captura o numero de atas complementares que o pregao possui
}
