% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numpreg.R
\name{get_numpreg}
\alias{get_numpreg}
\title{Captura numero do pregao}
\usage{
get_numpreg(html_doc, possui_ata_complementar = NA)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{possui_ata_complementar}{boolean indicando se o pregao possui ata complementar
complementar ou nao; se NA, a funcao ira verificar no html da ata}
}
\value{
String de 10 digitos com o numero do pregao
}
\description{
Captura o numero do pregao a partir do numero exibido no canto superior direito da ata. O numero do pregao corresponde ao numero entre o primeiro e segundo pontos, com zero-padding para completar 10 digitos.
}
