% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orgao.R
\name{get_orgao}
\alias{get_orgao}
\title{Captura nome do orgao}
\usage{
get_orgao(html_doc, possui_ata_complementar = NA)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{possui_ata_complementar}{boolean indicando se o pregao possui ata
complementar complementar ou nao; se NA, a funcao ira verificar no html da
ata}
}
\value{
string com o nome do orgao. Caso mais de uma linha de informacao sobre
  o orgao esteja disponivel na ata, as linhas sao separadas por
  ponto-e-virgula.
}
\description{
Captura nome do orgao, conforme exibido no inicio da ata, em posicao
centralizada.
}
