% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_propostas.R
\name{get_propostas}
\alias{get_propostas}
\title{Captura propostas do item}
\usage{
get_propostas(html_doc, item_id, quietly = TRUE)
}
\arguments{
\item{html_doc}{documento xml referente a ata de um pregao eletronico}

\item{item_id}{string com o numero do item}
}
\value{
dataframe com dados sobre as propostas
}
\description{
Captura as propostas recebidas para o item
}
