% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantidade.R
\name{get_quantidade}
\alias{get_quantidade}
\title{Extrai dado de quantidade a partir do texto do campo 'resultado'}
\usage{
get_quantidade(resultado)
}
\arguments{
\item{resultado}{character string com o resultado de um item, tal como disponivel no comeco das atas}
}
\value{
character
}
\description{
Extrai dado de quantidade a partir do texto do campo 'resultado'
}
\seealso{
para capturar texto de resultado a partir do html, ver get_item_info e get_item_info_chr
}
