% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valor_com_desconto.R
\name{get_valor_com_desconto}
\alias{get_valor_com_desconto}
\title{Extrai valor de melhor lance com desconto}
\usage{
get_valor_com_desconto(resultado)
}
\arguments{
\item{resultado}{character string com o resultado de um item, tal como disponivel no comeco das atas}
}
\value{
numeric
}
\description{
Extrai valor de melhor lance com desconto
}
\seealso{
para capturar texto de resultado a partir do html, ver get_item_info e get_item_info_chr
}
