% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valor_com_desconto_negociado.R
\name{get_valor_com_desconto_negociado}
\alias{get_valor_com_desconto_negociado}
\title{Extrai valor com desconto negociado}
\usage{
get_valor_com_desconto_negociado(resultado)
}
\arguments{
\item{resultado}{character string com o resultado de um item, tal como disponivel no comeco das atas}
}
\value{
numeric
}
\description{
Extrai valor com desconto negociado
}
\seealso{
para capturar texto de resultado a partir do html, ver get_item_info e get_item_info_chr
}
