% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vencedor.R
\name{get_vencedor}
\alias{get_vencedor}
\title{Extrai o nome do vencedor do pregao a partir do texto do campo 'resultado'}
\usage{
get_vencedor(resultado)
}
\arguments{
\item{resultado}{character string com o resultado de um item, tal como disponivel no comeco das atas}
}
\value{
character string com o nome do vencedor
}
\description{
Extrai o nome do vencedor do pregao a partir do texto do campo 'resultado'
}
