% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpar_valores.R
\name{limpar_valores}
\alias{limpar_valores}
\title{Converter valores disponiveis no historico de lances para formato numerico}
\usage{
limpar_valores(string)
}
\arguments{
\item{string}{character string com o valor do lance, tal como disponivel nas atas}
}
\value{
numeric
}
\description{
Converter valores disponiveis no historico de lances para formato numerico
}
