% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ata.R
\name{parse_ata}
\alias{parse_ata}
\title{Funcao para extrair dados do codigo html de uma ata}
\usage{
parse_ata(ata_html_file)
}
\arguments{
\item{ata_html_file}{caminho do arquivo html da ata cujas informacoes serao extraidas}
}
\value{
dataframe com os dados da ata
}
\description{
O principal uso desta funcao se da no contexto das funcoes parsePar e parseSeq, que capturam dados de varias atas de forma automatizada
}
