% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_resultado.R
\name{parse_resultado}
\alias{parse_resultado}
\title{Extrai diversas informacoes disponiveis no campo 'resultado'}
\usage{
parse_resultado(resultado, info = c("vencedor", "melhor_lance",
  "valor_negociado", "quantidade", "desconto", "desconto_negociado",
  "valor_com_desconto", "valor_com_desconto_negociado"))
}
\arguments{
\item{resultado}{character string com o resultado de um item, tal como disponivel no comeco das atas}

\item{info}{vetor com as informacoes desejadas}
}
\value{
Caso apenas uma informacao seja incluida em 'info', a funcao retorna um vetor numerico ou de caracteres . Caso mais informacoes sejam solicitadas, a funcao retorna um dataframe.
}
\description{
Extrai diversas informacoes disponiveis no campo 'resultado'
}
