% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePar.R
\name{parsePar}
\alias{parsePar}
\title{Parsing do html das atas usando computacao em paralelo}
\usage{
parsePar(atas_files, nworkers)
}
\arguments{
\item{atas_files}{vetor com os caminhos dos arquivos html das atas}

\item{nworkers}{numero de workers (nucleos) entre os quais o parsing das atas sera dividido; recomenda-se que seja no maximo igual ao numero de nucleos fisicos}
}
\value{
dataframe com os dados das atas
}
\description{
Funcao para extrair dados de um conjunto de atas (em formato html) usando computacao em paralelo para maior velocidade.
}
