% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSeq.R
\name{parseSeq}
\alias{parseSeq}
\title{Parsing do html das atas (sequencial)}
\usage{
parseSeq(atas_files)
}
\arguments{
\item{atas_files}{vetor com os caminhos dos arquivos html das atas}
}
\value{
dataframe com os dados das atas
}
\description{
Funcao que implementa um for loop para capturar os dados das atas (formato html) em um dataframe
}
