% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possui_ata_complementar.R
\name{possui_ata_complementar}
\alias{possui_ata_complementar}
\title{Verifica se o pregao da ata passada como argumento possui atas complementares ou nao.}
\usage{
possui_ata_complementar(html_doc)
}
\arguments{
\item{html_doc}{um documento xml referente a ata de um pregao eletronico}
}
\value{
Retorna TRUE se o pregao possui ata complementar e FALSE caso contrario.
}
\description{
Importante pois a existencia de atas complementares muda a estrutura do codigo html.
}
