% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_item_nodes.R
\name{select_item_nodes}
\alias{select_item_nodes}
\title{Selecionar item_nodes de interesse com base em um vetor de indices para acelerar o parsing}
\usage{
select_item_nodes(id_item_nodes, indices)
}
\arguments{
\item{id_item_nodes}{lista de nodes dos itens, obtida a partir de 'get_info_nodes'}

\item{indices}{vetor numerico com os indices dos itens de interesse}
}
\value{
lista de nodes dos itens de interesse
}
\description{
Selecionar item_nodes de interesse com base em um vetor de indices para acelerar o parsing
}
